Action Domination! Client/TEST/tool enhancements
Any time the game performs an action it goes through
the Action class now! (hooray for triggering!)

The client now supports a "run a list of commands" mode.
This mode logs the output to a file.

postprompt tool now has a "test" command to just do a simple
smoke test against the host.

Enhancements to diagram reader the and the option parser.

Updating diagram as I go.