Began work on revamping the gameplay screen.
Added a felt border as a play surface, which will likely be improved later.
Added classes to support it.
Replaced the older style card backs with new ones in the pixel style.
(Added the on-click rock border image, but it is not currently in use.)
Started to detangle the Javascript files supporting gameplay.  These will need
to be updated to add appropriate classes to generated divs and spans; right now
they all take on the class "player_status", which is incorrect.